Moved addStyleClassToVisibleFields into utils to be re-used
