+++ admin-base/ui.apps/src/main/content/jcr_root/apps/field/collection/template.vue
@@ -87,6 +87,8 @@
 </template>
 
 <script>
+import {addStyleClassToVisibleFields} from '../../../../../js/utils'
+
 export default {
   mixins: [VueFormGenerator.abstractField],
   beforeMount() {
@@ -108,71 +110,6 @@ export default {
     }
   },
   methods: {
-    addStyleClassToVisibleFields(model, schema) {
-      const safeEval = (expr) => {
-        if (typeof expr !== 'string') return true;
-      
-        const fixed = expr
-          .replace(/\band\b/g, '&&')
-          .replace(/\bor\b/g, '||')
-          .replace(/\bnot\b/g, '!')
-      
-        try {
-          return Function("model", '"use strict"; return (' + fixed + ')')(model);
-        } catch (e) {
-          console.warn('Error evaluating visible:', expr, e);
-          return true;
-        }
-      };
-
-      const deepClone = (obj) => {
-        if (obj === null || typeof obj !== 'object') return obj;
-      
-        // Handle arrays
-        if (Array.isArray(obj)) {
-          return obj.map(item => deepClone(item));
-        }
-      
-        // Handle objects
-        const clonedObj = {};
-        for (const key in obj) {
-          // Only copy own properties
-          if (Object.hasOwn(obj, key)) {
-            clonedObj[key] = deepClone(obj[key]);
-          }
-        }
-        return clonedObj;
-      }
-
-      const clone = deepClone(schema);
-          
-          const process = (obj) => {
-            if (!obj) return;
-            
-            // Process groups
-            if (obj.groups) {
-              obj.groups.forEach(group => process(group));
-            }
-            
-            // Process fields
-            if (obj.fields) {
-              obj.fields.forEach(field => {
-                // Handles nested fields
-                if (field.fields) {
-                  process(field);
-                }
-
-                if (field.visible !== undefined) {                
-                  const result = safeEval(field.visible);        
-                  field.styleClasses = result ? '' : 'hidden';
-                }
-              })
-            }
-          }
-          
-      process(clone);
-      return clone;
-        },
     getSchemaForIndex(schema, index) {
       const newSchema = JSON.parse(JSON.stringify(schema))
       newSchema.fields[0].model = '' + index
@@ -202,7 +139,7 @@ export default {
       return parseInt(index) + 1
     },
     prepSchema(model, schema) {
-      const clonedSchema = this.addStyleClassToVisibleFields(model, schema);
+      const clonedSchema = addStyleClassToVisibleFields(schema, model);
     
       for (let i = 0; i < clonedSchema.fields.length; i++) {
         const field = clonedSchema.fields[i].model;
